export parser tests to its hown file
